fix bug

+++ app/Http/Controllers/ordersController.php
@@ -77,8 +77,7 @@ class ordersController extends Controller
                     // return response()->json($getProduct->quantity);
                 } else {
                     // $getProduct->quantity = $getProduct->quantity - $getProduct->Unit_Pieces *  $order_items->product_quantity;
-                    $getProduct->quantity = $getProduct->quantity - (float)$getProduct->Unit_Pieces * (float)$order_items->product_quantity;
-
+                    $getProduct->quantity = $getProduct->quantity - (int)$getProduct->Unit_Pieces * (int)$order_items->product_quantity;
                     $getProduct->update();
                 }
             };
